Split off tag tests from main docker testbench
