+++ src/test/resources/testbash/auxiliary_tag/check_bash_version.sh
