+++ src/test/resources/testbash/auxiliary_tag/config_bash_tag.vsh.yaml
@@ -1,5 +1,5 @@
 functionality:
-  name: testbash
+  name: testbash_tag
   description: |
     Prints out version of bash.
   resources:
