+++ src/test/scala/com/dataintuitive/viash/MainBuildDockerSuite.scala
@@ -24,10 +24,6 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
   private val executable = Paths.get(tempFolStr, functionality.name).toFile
   private val execPathInDocker = Paths.get("/viash_automount", executable.getPath).toFile.toString
 
-  private val configBashTagFile = getClass.getResource(s"/testbash/config_bash_tag.vsh.yaml").getPath
-  private val functionalityBashTag = Config.read(configBashTagFile, modifyFun = false).functionality
-  private val executableBashTagFile = Paths.get(tempFolStr, functionalityBashTag.name).toFile
-
   //<editor-fold desc="Test benches to build a generic script and run various commands to see if the functionality is correct">
   // convert testbash
   test("viash can create an executable") {
@@ -185,170 +181,6 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
     assert(regexPlatform.findFirstIn(testText).isDefined, testText)
   }
   //</editor-fold>
-  //<editor-fold desc="Test benches to check building with tags, versions & registries">
-  test("Get tagged version of a docker image for bash 5.0", DockerTest) {
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", "testtag1",
-      "-o", tempFolStr,
-      configBashTagFile
-    )
-
-    assert(executableBashTagFile.exists)
-    assert(executableBashTagFile.canExecute)
-
-    // create the docker image
-    val out = Exec.run2(
-      Seq(executableBashTagFile.toString, "---setup", "build")
-    )
-    assert(out.exitValue == 0)
-
-    // run the script
-    val stdout = Exec.run(Seq(executableBashTagFile.toString))
-    assert(stdout.contains("GNU bash, version 5.0"))
-
-    // check whether the internal docker is correct
-    val dockerout = Exec.run(Seq(executableBashTagFile.toString, "---dockerfile"))
-    // we expect something basic like
-    // FROM bash:5.0
-    //
-    // RUN :
-    // Allow for extra spaces just in case the format changes slightly format-wise but without functional differences
-    val regex = """^FROM bash:5\.0[\r\n\s]*RUN\s+:\s*$""".r
-    assert(regex.findFirstIn(dockerout).isDefined, regex.toString)
-  }
-
-  test("Get tagged version of a docker image for bash 3.2", DockerTest) {
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", "testtag2",
-      "-o", tempFolStr,
-      configBashTagFile
-    )
-
-    assert(executableBashTagFile.exists)
-    assert(executableBashTagFile.canExecute)
-
-    // create the docker image
-    val out = Exec.run2(
-      Seq(executableBashTagFile.toString, "---setup", "build")
-    )
-    assert(out.exitValue == 0)
-
-    // run the script
-    val stdout = Exec.run(Seq(executableBashTagFile.toString))
-    assert(stdout.contains("GNU bash, version 3.2"))
-
-    // check whether the internal docker is correct
-    val dockerout = Exec.run(Seq(executableBashTagFile.toString, "---dockerfile"))
-    // we expect something basic like
-    // FROM bash:3.2
-    //
-    // RUN :
-    // Allow for extra spaces just in case the format changes slightly format-wise but without functional differences
-    val regex = """^FROM bash:3\.2[\r\n\s]*RUN\s+:\s*$""".r
-    assert(regex.findFirstIn(dockerout).isDefined, regex.toString)
-  }
-
-  test("Check whether target image name is well formed without target_image, version or registry", DockerTest) {
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", "testtargetimage1",
-      "-o", tempFolStr,
-      configBashTagFile
-    )
-
-    assert(executableBashTagFile.exists)
-    assert(executableBashTagFile.canExecute)
-
-    // create the docker image
-    val out = Exec.run2(
-      Seq(executableBashTagFile.toString, "---setup", "build")
-    )
-    assert(out.exitValue == 0)
-
-    // run the script
-    val stdout = Exec.run(Seq(executableBashTagFile.toString))
-    assert(stdout.contains("GNU bash, version 5.0"))
-
-    // check whether script is using the expected docker image
-    val contentSource = Source.fromFile(executableBashTagFile)
-    val content = try {
-      contentSource.getLines().toList
-    } finally {
-      contentSource.close()
-    }
-    assert(content.exists(_.matches("cat << VIASHEOF \\| eval docker run .* testbash:latest")))
-  }
-
-  test("Check whether target image name is well formed with target_image, version, and registry", DockerTest) {
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", "testtargetimage2",
-      "-o", tempFolStr,
-      configBashTagFile
-    )
-
-    assert(executableBashTagFile.exists)
-    assert(executableBashTagFile.canExecute)
-
-    // create the docker image
-    val out = Exec.run2(
-      Seq(executableBashTagFile.toString, "---setup", "build")
-    )
-    assert(out.exitValue == 0)
-
-    // run the script
-    val stdout = Exec.run(Seq(executableBashTagFile.toString))
-    assert(stdout.contains("GNU bash, version 5.0"))
-
-    // check whether script is using the expected docker image
-    val contentSource = Source.fromFile(executableBashTagFile)
-    val content = try {
-      contentSource.getLines().toList
-    } finally {
-      contentSource.close()
-    }
-    assert(content.exists(_.matches("cat << VIASHEOF \\| eval docker run .* foo.io/bar:0\\.0\\.1")))
-  }
-
-  test("Check whether target image name is well formed with target_image, target_tag", DockerTest) {
-    // prepare the environment
-    TestHelper.testMain(
-      "build",
-      "-p", "testtargetimage3",
-      "-o", tempFolStr,
-      configBashTagFile
-    )
-
-    assert(executableBashTagFile.exists)
-    assert(executableBashTagFile.canExecute)
-
-    // create the docker image
-    val out = Exec.run2(
-      Seq(executableBashTagFile.toString, "---setup", "build")
-    )
-    assert(out.exitValue == 0)
-
-    // run the script
-    val stdout = Exec.run(Seq(executableBashTagFile.toString))
-    assert(stdout.contains("GNU bash, version 3.2"))
-
-    // check whether script is using the expected docker image
-    val contentSource = Source.fromFile(executableBashTagFile)
-    val content = try {
-      contentSource.getLines().toList
-    } finally {
-      contentSource.close()
-    }
-    assert(content.exists(_.matches("cat << VIASHEOF \\| eval docker run .* bar:0\\.0\\.2")))
-  }
-
-  //</editor-fold>
   //<editor-fold desc="Test benches to check building with --meta flag">
   test("Get meta data of a docker", DockerTest) {
     // Create temporary folder to copy the files to so we can do a git init in that folder
@@ -378,8 +210,8 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
         configMetaFile
       )
 
-      assert(executableBashTagFile.exists)
-      assert(executableBashTagFile.canExecute)
+      assert(executable.exists)
+      assert(executable.canExecute)
 
       val viashVersion = com.dataintuitive.viash.Main.version
 
@@ -447,8 +279,8 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
         configMetaFile
       )
 
-      assert(executableBashTagFile.exists)
-      assert(executableBashTagFile.canExecute)
+      assert(executable.exists)
+      assert(executable.canExecute)
 
       val viashVersion = com.dataintuitive.viash.Main.version
 
@@ -507,8 +339,8 @@ class MainBuildDockerSuite extends FunSuite with BeforeAndAfterAll {
         configMetaFile
       )
 
-      assert(executableBashTagFile.exists)
-      assert(executableBashTagFile.canExecute)
+      assert(executable.exists)
+      assert(executable.canExecute)
 
       val viashVersion = com.dataintuitive.viash.Main.version
